requirements.txt
migra ~= 3.0.0
psycopg2-binary ~= 2.9.0
